Centralize File Extension Matcher
Remove the duplicate code that the file matcher logic uses and have it at the top of the entry file instead of declared inline.
Make sure that all regular expressions are Unicode compatible and that they are case insensitive matches.
This behavior that should be caught on the Windows platform, as Windows is case insensitive and this could cause a namespace collision if not insensitive matching.
While possible to have two different files with the same name only differing on linux and MacOS, this is generally agreed upon as a bad practice.