Modernize Output
Since Node 18 LTS is the earliest supported version of Node.JS, match the TS output with the maximum version of ECMAscript that is supported by Node.JS 18 LTS.
Disable declaration source maps since the source maps are already included with the project.
Disable downlevel iteration since ES2022 is the target and it supports modern iteration. This will improve performance.
Remove comments from the transpiled output since they aren't necessary in production, this will reduce bundle size of the final project.
Since the target is now ES2022, the lib config is not needed since the functionality is now included in the language output.