Update YAML Glob
Remove the standard wildcard YAML check as there are none on the root of the project.
Removed the GitHub prefix on the final YAML glob so that it checks for all YAML files, not just the ones in the GitHub folder since the intent was to check all YAML with prettier.