Deep Integrate Dependabot
Dependabot only scans for security risks by default.
This commit adds the ability for it to find package updates automatically.
It will create a PR which can be merged easily or even automatically if all the checks pass which allows for automated dependency management when there are no detected breaking changes.